+++ lib/helpers.js
@@ -10,6 +10,8 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const jsonld = require("jsonld");
 const isuri = require("isuri");
+const node_fetch_1 = require("node-fetch");
+const config_1 = require("./config");
 function iriify(str) {
     return `<${str}>`;
 }
@@ -55,3 +57,38 @@ exports.docToQuads = (doc) => __awaiter(this, void 0, void 0, function* () {
         });
     });
 });
+function getDoc(subject) {
+    return __awaiter(this, void 0, void 0, function* () {
+        console.log('Getting doc:', subject);
+        const query = `
+  var subject = ${JSON.stringify(encodeIRI(subject))};
+  g.V(subject)
+  	.OutPredicates()
+  	.ForEach(function mapPredicates(node) {
+        var predicate = node.id;
+        return g.V(subject)
+          .Out(predicate)
+          .ForEach(function emitObject(node) {
+            var object = node.id;
+            g.Emit({
+              subject: subject,
+              predicate: predicate,
+              object: object
+            });
+          });
+      })`;
+        const url = `${config_1.CAYLEY_ADDRESS}/api/v1/query/gizmo`;
+        return node_fetch_1.default(url, {
+            method: 'post',
+            body: query
+        })
+            .then(res => {
+            return res.json();
+        })
+            .then(({ result }) => {
+            console.log('Returning resulting quads:', result);
+            return result;
+        });
+    });
+}
+exports.getDoc = getDoc;
