+++ lib/index.js
@@ -9,64 +9,15 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const memux_1 = require("memux");
-const node_fetch_1 = require("node-fetch");
+const feedbackfruits_knowledge_engine_1 = require("feedbackfruits-knowledge-engine");
 const Cayley = require('cayley');
-const { Observable: { empty } } = require('rxjs');
-const PQueue = require('p-queue');
 const jsonld = require('jsonld');
 const Config = require("./config");
 const helpers_1 = require("./helpers");
 const cayley = Cayley(Config.CAYLEY_ADDRESS);
-function getDoc(subject) {
+function init({ name }) {
     return __awaiter(this, void 0, void 0, function* () {
-        console.log('Getting doc:', subject);
-        const g = cayley.graph;
-        const query = `
-  var subject = ${JSON.stringify(helpers_1.encodeIRI(subject))};
-  g.V(subject)
-  	.OutPredicates()
-  	.ForEach(function mapPredicates(node) {
-        var predicate = node.id;
-        return g.V(subject)
-          .Out(predicate)
-          .ForEach(function emitObject(node) {
-            var object = node.id;
-            g.Emit({
-              subject: subject,
-              predicate: predicate,
-              object: object
-            });
-          });
-      })`;
-        const url = `${Config.CAYLEY_ADDRESS}/api/v1/query/gizmo`;
-        return node_fetch_1.default(url, {
-            method: 'post',
-            body: query
-        })
-            .then(res => {
-            return res.json();
-        })
-            .then(({ result }) => {
-            console.log('Returning resulting quads:', result);
-            return result;
-        });
-    });
-}
-function init({ name, url, input, output }) {
-    return __awaiter(this, void 0, void 0, function* () {
-        const ssl = {
-            key: Config.KAFKA_PRIVATE_KEY,
-            cert: Config.KAFKA_CERT,
-            ca: Config.KAFKA_CA,
-        };
-        const send = yield memux_1.createSend({
-            name,
-            url,
-            topic: output,
-            concurrency: Config.CONCURRENCY,
-            ssl
-        });
-        const receive = (operation) => __awaiter(this, void 0, void 0, function* () {
+        const receive = (send) => (operation) => __awaiter(this, void 0, void 0, function* () {
             if (!memux_1.isOperation(operation))
                 throw new Error();
             const { action, data } = operation;
@@ -80,7 +31,7 @@ function init({ name, url, input, output }) {
                     if (error != null)
                         return reject(error);
                     if (response.statusCode >= 200 && response.statusCode < 400) {
-                        return Promise.all(helpers_1.quadsToDocs(yield getDoc(data['@id'])).map(data => send({ action, key: data['@id'], data }))).then(() => resolve());
+                        return Promise.all(helpers_1.quadsToDocs(yield helpers_1.getDoc(data['@id'])).map(data => send({ action, key: data['@id'], data }))).then(() => resolve());
                     }
                     if (response.statusCode === 400) {
                         if ((body.error || body).match(/quad exists/))
@@ -92,23 +43,17 @@ function init({ name, url, input, output }) {
                 }));
             });
         });
-        return memux_1.createReceive({
+        return feedbackfruits_knowledge_engine_1.Annotator({
             name,
-            url,
-            topic: input,
             receive,
-            ssl
+            customConfig: Config
         });
     });
 }
 exports.default = init;
 if (require.main === module) {
     console.log("Running as script.");
-    console.log('Initializing with config:', Config);
     init({
         name: Config.NAME,
-        url: Config.KAFKA_ADDRESS,
-        input: Config.INPUT_TOPIC,
-        output: Config.OUTPUT_TOPIC,
     }).catch(console.error);
 }
